handleTurns - playG PGraphic - activeTank
The function previously called swapTurn has been renamed to handleTurns. As well, the function has been fixed such that the active tank display in the player data bar now correctly displays the active tank for each player.

The player data bar was previously a function that drew the bar, this has been moved to be a part of the playG PGraphic for optimization sake. Note backdrop Pgraphic has been renamed to playG.