chatBox class. First working chatBox object.
A ChatBox class was created. Its main function is to display as much of its string list (from end to start) as the height of the chatBox allows.

The chatBox also has the potential to scroll up and down in terms of its viewing location if the full log of strings is longer than the chatbox height. 

Note next commit the two buttons or the slidebar which will control this needs to be implemented.
As well, a textBox to input strings is needed.