textButton(button) Class created and implemented
Pressing Enter on any page will enable that page's chatBox and clicking a chatBox enables it as well. 
Once enabled, the textButton updates it's string as the user types. After a string has been inputted and the user hit's enter, the string then gets added to that textButton's associated chatBox.

Main screen has the first implementation of a working chatBox and textButton connected.