ValueBar class. FirePower and FireAngle bars.
ValueBar class allows for the modification of a value via the bar. Whenever the user clicks along a bar, the value the bar stores will change, as well as the value indicator on the bar.

These value bars will be used to control the fire power and fire angle of players when firing.